feat: 수동 로또와 자동 로또를 합친 등수 계산 기능 구현

+++ src/main/kotlin/lotto/domain/LottoResult.kt
@@ -13,8 +13,8 @@ class LottoResult(
     }
 
     fun calculateProfitRate(price: Int): Double {
-        val totalWinningMoney = getRanks().sumOf { it.winningMoney }
-        val totalCost = Purchase(price).getPrice()
+        val totalWinningMoney = getRanks().sumOf { it.winningMoney.toLong() }
+        val totalCost = Purchase(price).getPrice().toLong()
         return totalWinningMoney.toDouble() / totalCost
     }
 }
